# docs/openapi/pn-apikey-manager-internal-v1.yaml
openapi: 3.0.3
info:
  termsOfService: https://termofservice.it
  x-api-id: api-internal-web-pa
  title: 'Piattaforma Notifiche: API per il Front End delle PA'
  x-summary: 'Piattaforma Notifiche: API per il Front End delle PA'
  version: '1.0.0'
  description: >-
    ## Abstract
      API utilizzate dal portale di Piattaforma Notifiche dedicato alle Pubbliche Amministrazioni
  contact:
    email: pn@pagopa.it
  license:
    name: Licenza di PN
    url: 'https://da-definire/'
servers:
  - url: https://webapi.pn.pagopa.it
    description: Ambiente di produzione
  - url: https://webapi.uat.pn.pagopa.it
    description: Ambiente di test
  - url: https://webapi.dev.pn.pagopa.it
    description: Ambiente di sviluppo
tags:
  - name: ApiKeys
    description: >-
      Invocazioni utilizzabili dai mittenti per la gestione delle api keys


paths:
  "/api-keys":
    get:
      summary: Ricerca api keys
      description: >-
        servizio di consultazione della lista delle api keys
      tags:
        - ApiKeys
      operationId: getApiKeys
      parameters:                                           # NO EXTERNAL
        - $ref: '#/components/parameters/uidAuthFleet'      # NO EXTERNAL
        - $ref: '#/components/parameters/cxTypeAuthFleet'   # NO EXTERNAL
        - $ref: '#/components/parameters/cxIdAuthFleet'     # NO EXTERNAL
        - $ref: '#/components/parameters/cxGroupsAuthFleet' # NO EXTERNAL
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10

        - name: lastKey
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiKeysResponse'
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
    post:
      summary: Creazione api key
      description: >-
        servizio di creazione di un'api key
      tags:
        - ApiKeys
      operationId: newApiKey
      parameters:                                           # NO EXTERNAL
        - $ref: '#/components/parameters/uidAuthFleet'      # NO EXTERNAL
        - $ref: '#/components/parameters/cxTypeAuthFleet'   # NO EXTERNAL
        - $ref: '#/components/parameters/cxIdAuthFleet'     # NO EXTERNAL
        - $ref: '#/components/parameters/cxGroupsAuthFleet' # NO EXTERNAL
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestNewApiKey"
        required: true
      responses:
        '200':
          description: OK
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ResponseNewApiKey'
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'

  "/api-keys/{id}":
    delete:
      summary: Rimozione api key
      description: >-
        servizio di rimozione dell'api key
      tags:
        - ApiKeys
      operationId: deleteApiKeys
      parameters:
        - $ref: '#/components/parameters/uidAuthFleet' # NO EXTERNAL
        - $ref: '#/components/parameters/cxTypeAuthFleet' # NO EXTERNAL
        - $ref: '#/components/parameters/cxIdAuthFleet' # NO EXTERNAL
        - $ref: '#/components/parameters/cxGroupsAuthFleet' # NO EXTERNAL
        - $ref: '#/components/parameters/pathApiKey'
      responses:
        '200':
          description: OK
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: Wrong state transition (i.e. delete an enabled key)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'

  "/api-keys/{id}/status":
    put:
      summary: Cambia lo stato dell'api key
      description: >-
        servizio di cambio stato dell'api key
      tags:
        - ApiKeys
      operationId: changeStatusApiKey
      parameters:
        - $ref: '#/components/parameters/uidAuthFleet' # NO EXTERNAL
        - $ref: '#/components/parameters/cxTypeAuthFleet' # NO EXTERNAL
        - $ref: '#/components/parameters/cxIdAuthFleet' # NO EXTERNAL
        - $ref: '#/components/parameters/cxGroupsAuthFleet' # NO EXTERNAL
        - $ref: '#/components/parameters/pathApiKey'
        - $ref: '#/components/parameters/queryApiKey'
      responses:
        '200':
          description: OK
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: Wrong state transition (i.e. enable an enabled key)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'

components:
  parameters:
    ############################################################################################
    ###                     PARAMETRI DI AUTENTICAZIONE E AUTORIZZAZIONE                     ###
    ############################################################################################
    ### N.B. PRENDERLE DA AUTH FLEET ###
    cxTypeAuthFleet:                                                       # NO EXTERNAL
      $ref: './remote-refs.yaml#/components/parameters/cxTypeAuthFleet'    # NO EXTERNAL
    cxIdAuthFleet:                                                         # NO EXTERNAL
      $ref: './remote-refs.yaml#/components/parameters/cxIdAuthFleet'      # NO EXTERNAL
    cxGroupsAuthFleet:                                                     # NO EXTERNAL
      $ref: './remote-refs.yaml#/components/parameters/cxGroupsAuthFleet'  # NO EXTERNAL
    uidAuthFleet:                                                          # NO EXTERNAL
      $ref: './remote-refs.yaml#/components/parameters/uidAuthFleet'       # NO EXTERNAL

    ############################################################################################
    ###                        PARAMETRI RICERCA API KEY                                     ###
    ############################################################################################
    pathApiKey:
      description: >-
        Identificativo univoco dell'api key
      name: id
      in: path
      required: true
      schema:
        type: string
    queryApiKey:
      description: >-
        Action per il cambio stato di un'api key
      name: status
      in: query
      required: true
      schema:
        type: string
        enum:
          - 'BLOCK'
          - 'ENABLE'
          - 'ROTATE'
  schemas:
    Problem:
      $ref: './remote-refs.yaml#/components/schemas/Problem'

    ApiKeysResponse:
      title: Elenco di api keys
      description: >-
        Dto contenente la lista delle api keys associate ad un utente.
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ApiKeyRow'

    ApiKeyRow:
      type: object
      properties:
        id:
          type: string
          description: Id dell'api key
        name:
          type: string
          description: Nome dell'api key
        value:
          type: string
          description: Valore dell'api key
        lastUpdate:
          type: string
          description: Data ultima modifica
          format: date-time
        groups:
          type: array
          description: Gruppi a cui appartiene l'api key
          items:
            type: string
        status:
          $ref: '#/components/schemas/ApiKeyStatus'
        statusHistory:
          type: array
          description: Storico degli stati dell'api key
          items:
            $ref: '#/components/schemas/ApiKeyStatusHistory'

    ApiKeyStatus:
      description: Stato dell'api key
      type: string
      enum: ['CREATED', 'ENABLED', 'BLOCKED', 'ROTATED']

    ApiKeyStatusHistory:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/ApiKeyStatus'
        date:
          type: string
          description: data a cui corrisponde il cambio di stato
          format: date-time
        changedByDenomination:
          type: string
          description: nome dell'utente che ha effettuato il cambio di stato

    RequestNewApiKey:
      title: Request creazione api key
      description: >-
        Request per la creazione di una nuova api key
      type: object
      required:
        - name
        - groups
      properties:
        name:
          type: string
          description: nome dell'api key
        groups:
          type: array
          description: Gruppi a cui appartiene l'api key (indicare l'id dei gruppi)
          items:
            type: string

    ResponseNewApiKey:
      title: Response creazione api key
      description: >-
        Response per la creazione di una nuova api key
      type: object
      required:
        - id
        - api-key
      properties:
        id:
          type: string
          description: id dell'api key appena generata
        api-key:
          type: string
          description: Valore dell'api key appena generata

#  securitySchemes:                                      # ONLY EXTERNAL
#    bearerAuth:                                         # ONLY EXTERNAL
#      type: http                                        # ONLY EXTERNAL
#      scheme: bearer                                    # ONLY EXTERNAL
#      bearerFormat: JWT                                 # ONLY EXTERNAL
#security:                                               # ONLY EXTERNAL
# - bearerAuth: []                                       # ONLY EXTERNAL